refactor(model-router): 循環依存の解消と windowSize エッジケース対応
AI レビュー指摘事項への対応:

1. 循環依存解消
   - classifier.ts と session-store.ts が相互に import type していた問題を解消
   - 共通ドメイン型（ClassificationResult, ClassificationReason,
     ClassificationDetail, TurnRecord, SessionContext）を types.ts に集約
   - 後方互換のため両ファイルから型を re-export

2. shouldStickyDefault の windowSize <= 0 ガード
   - JS の Array.prototype.slice(-0) === slice(0) の罠を回避
   - windowSize: 0 設定時に意図せず全ターンを参照する挙動を防止
   - windowSize=0 / -1 のテストケースを追加
